+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # OldNews ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added a busy dialog when we're waiting for TheOldReader to add a new
+  subscription. ([#99](https://github.com/davep/oldnews/pull/99))
+
 ## v0.7.0
 
 **Released: 2026-01-28**